requirements.txt
easydict==1.9
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
protobuf==3.11.2
grpcio==1.32
grpcio-tools==1.32
grpcio-health-checking==1.32
segmentation_models==1.0.1
onnx==1.6.0
onnxconverter-common==1.7.0
onnxruntime==1.5.1
contexttimer==0.3

mypy-protobuf
watchdog
argh
pyyaml
